requirements.txt
PyQt6
Pillow
numpy
opencv-python
imageio
librosa
moviepy
